requirements.txt
requests
pandas
selenium
beautifulsoup4
python-dotenv
lxml
html2text
markdownify
pyyaml
fastapi
uvicorn
pydantic
openai